fix: 자동완성 검색 두 번 클릭하게 되는 버그 해결
