+++ client/src/hooks/fetch/useSearchDocumentByQuery.ts
@@ -5,18 +5,18 @@ import useDebounce from '../useDebounce';
 import {useFetch} from '@hooks/useFetch';
 import {requestGet} from '@apis/http';
 
+const getSearchDocument = async (query: string) => {
+  const response = requestGet<string[]>({
+    baseUrl: process.env.NEXT_PUBLIC_BASE_URL,
+    endpoint: `/api/get-search-document?referQuery=${query}`,
+  });
+
+  return response;
+};
+
 const useSearchDocumentByQuery = (query: string) => {
-  const getSearchDocument = async (query: string) => {
-    const response = requestGet<string[]>({
-      baseUrl: process.env.NEXT_PUBLIC_BASE_URL,
-      endpoint: `/api/get-search-document?referQuery=${query}`,
-    });
-
-    return response;
-  };
-
   const searchDocumentByQuery = useCallback(() => getSearchDocument(query), [query]);
-  const {data, refetch, setData} = useFetch(searchDocumentByQuery, {enabled: true});
+  const {data, refetch} = useFetch(searchDocumentByQuery, {enabled: true});
 
   const searchDocumentsIfValid = useCallback(() => {
     if (query.trim() !== '' && /^[가-힣()0-9]*$/.test(query)) refetch();
@@ -25,9 +25,8 @@ const useSearchDocumentByQuery = (query: string) => {
   const debouncedSearchDocuments = useDebounce(searchDocumentsIfValid, 200);
 
   useEffect(() => {
-    setData(null);
     debouncedSearchDocuments();
-  }, [debouncedSearchDocuments, query, setData]);
+  }, [debouncedSearchDocuments, query]);
 
   return {
     titles: data ?? [],
